tribeux-domtree/pyproject.toml
[project]
name = "tribeux-domtree"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.40",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "pillow>=10",
    "pydantic>=2.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
